Reorganized components and fixed Toastify issue
