+++ src/app/admin/layout.tsx
@@ -1,7 +1,6 @@
 import type { Metadata } from "next";
-import "react-toastify/dist/ReactToastify.css";
 
-import SideNavBar from "./components/SideNavBar";
+import SideNavBar from "@/components/admin/SideNavBar";
 
 export const metadata: Metadata = {
   title: "Hack the Change - Admin",
